Completely rework event system
Now uses reflection to find event handling, data getting, and event filtering functions defined on the class that implements MultiEventHandler.

The old system was replaced because it was hard to cleanly override event handlers from subclasses.

Event handlers can access data by simply adding it as a parameter to the function. If the class implements DataStore then it can also take the store's DATA directly (handler is wrapped in DataStore.withData)